Quad仕様を外部ファイルに分離
