+++ src/process/onetime/season_1673282_specialized_quad_process.ts
@@ -14,194 +14,12 @@ import { MoveToTargetTask } from "v5_object_task/creep_task/combined_task/move_t
 import { BoostApiWrapper } from "v5_object_task/creep_task/api_wrapper/boost_api_wrapper"
 import { OperatingSystem } from "os/os"
 import { Quad, QuadAttackTargetType, QuadState } from "./season_1673282_specialized_quad"
+import { QuadSpec, QuadType } from "./season_1673282_specialized_quad_spec"
 import { CreepName } from "prototype/creep"
 import { GameConstants } from "utility/constants"
 import { boostableCreepBody } from "utility/resource"
 
 type AttackTarget = AnyCreep | AnyStructure
-type CreepBodySpec = {
-  roles: CreepRole[]
-  body: BodyPartConstant[]
-}
-
-export const season1673282SpecializedQuadProcessCreepType = [
-  "test-dismantler",
-  "test-attacker",
-  "invader-core-attacker",
-  "tier0-d100-attacker",
-  "tier0-d450",
-  "tire0-d360-dismantler",
-  "no-defence-3tower",
-  "tier3-d2000-dismantler-swamp",
-] as const
-type Season1673282SpecializedQuadProcessCreepType = typeof season1673282SpecializedQuadProcessCreepType[number]
-
-export const isSeason1673282SpecializedQuadProcessCreepType = (obj: string): obj is Season1673282SpecializedQuadProcessCreepType => {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  return season1673282SpecializedQuadProcessCreepType.includes(obj as any)
-}
-
-const noBoosts: MineralBoostConstant[] = [
-]
-
-const testHealerSpec: CreepBodySpec = {
-  roles: [CreepRole.RangedAttacker, CreepRole.Healer, CreepRole.Mover],
-  body: [RANGED_ATTACK, MOVE, MOVE, HEAL],
-}
-const testDismantlerSpec: CreepBodySpec = {
-  roles: [CreepRole.Worker, CreepRole.Healer, CreepRole.Mover],
-  body: [WORK, MOVE, MOVE, HEAL],
-}
-const testAttackerSpec: CreepBodySpec = {
-  roles: [CreepRole.Worker, CreepRole.Healer, CreepRole.Mover],
-  body: [ATTACK, MOVE, MOVE, HEAL],
-}
-
-const invaderCoreAttackerAttacker: CreepBodySpec = {
-  roles: [CreepRole.Attacker, CreepRole.Mover],
-  body: [
-    ATTACK, ATTACK, ATTACK, ATTACK, ATTACK,
-    MOVE, MOVE, MOVE, MOVE, MOVE,
-  ],
-}
-const invaderCoreAttackerHealer: CreepBodySpec = {
-  roles: [CreepRole.RangedAttacker, CreepRole.Healer, CreepRole.Mover],
-  body: [
-    RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK,
-    MOVE, MOVE, MOVE, MOVE, MOVE,
-    HEAL, HEAL,
-  ],
-}
-
-const tier0h3HealerSpec: CreepBodySpec = {
-  roles: [CreepRole.RangedAttacker, CreepRole.Healer, CreepRole.Mover],
-  body: [
-    RANGED_ATTACK, MOVE, RANGED_ATTACK, MOVE, RANGED_ATTACK, MOVE, RANGED_ATTACK, MOVE, RANGED_ATTACK, MOVE,
-    MOVE, MOVE, MOVE,
-    HEAL, HEAL, HEAL,
-  ]
-}
-const tier0AttackerSpec: CreepBodySpec = {
-  roles: [CreepRole.Attacker, CreepRole.Healer, CreepRole.Mover],
-  body: [
-    RANGED_ATTACK, MOVE, RANGED_ATTACK, MOVE, RANGED_ATTACK, MOVE, RANGED_ATTACK, MOVE, RANGED_ATTACK, MOVE,
-    ATTACK, MOVE, ATTACK, MOVE, ATTACK, MOVE,
-  ]
-}
-
-const tire0h10HealerSpec: CreepBodySpec = {
-  roles: [CreepRole.RangedAttacker, CreepRole.Healer, CreepRole.Mover],
-  body: [
-    RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK,
-    RANGED_ATTACK, MOVE, RANGED_ATTACK, MOVE, RANGED_ATTACK, MOVE, RANGED_ATTACK, MOVE, RANGED_ATTACK, MOVE,
-    RANGED_ATTACK, MOVE, RANGED_ATTACK, MOVE, RANGED_ATTACK, MOVE, RANGED_ATTACK, MOVE, RANGED_ATTACK, MOVE,
-    MOVE, MOVE, MOVE, MOVE, MOVE,
-    MOVE, MOVE, MOVE, MOVE, MOVE,
-    MOVE, MOVE, MOVE, MOVE,
-    HEAL, HEAL, HEAL, HEAL, HEAL,
-    HEAL, HEAL, HEAL, HEAL,
-    MOVE, HEAL,
-  ]
-}
-const tire0DismantlerSpec: CreepBodySpec = {
-  roles: [CreepRole.Worker, CreepRole.Mover],
-  body: [
-    WORK, WORK, WORK, WORK,
-    WORK, MOVE, WORK, MOVE, WORK, MOVE, WORK, MOVE, WORK, MOVE,
-    WORK, MOVE, WORK, MOVE, WORK, MOVE, WORK, MOVE, WORK, MOVE,
-    WORK, MOVE, WORK, MOVE, WORK, MOVE, WORK, MOVE, WORK, MOVE,
-    WORK, MOVE, WORK, MOVE, WORK, MOVE, WORK, MOVE, WORK, MOVE,
-    WORK, MOVE,
-    MOVE, MOVE, MOVE, MOVE,
-  ]
-}
-
-// const noDefence3TowerAttackerSpec: CreepBodySpec = {
-//   roles: [CreepRole.RangedAttacker, CreepRole.Healer, CreepRole.Mover],
-//   body: [
-//     TOUGH, TOUGH, TOUGH, TOUGH, TOUGH,
-//     TOUGH,
-//     RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK,
-//     RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK,
-//     MOVE, MOVE, MOVE, MOVE, MOVE,
-//     RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK,
-//     RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK,
-//     RANGED_ATTACK, RANGED_ATTACK,
-//     MOVE, MOVE, MOVE, MOVE, MOVE,
-//     MOVE, MOVE, MOVE, MOVE, MOVE,
-//     MOVE, MOVE,
-//     HEAL, HEAL, HEAL, HEAL, HEAL,
-//   ]
-// }
-const noDefence3TowerAttackerSpec: CreepBodySpec = {  // for RCL7
-  roles: [CreepRole.RangedAttacker, CreepRole.Healer, CreepRole.Mover],
-  body: [
-    TOUGH, TOUGH, TOUGH, TOUGH, TOUGH,
-    TOUGH,
-    RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK,
-    RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK,
-    MOVE, MOVE, MOVE, MOVE,
-    RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK,
-    RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK,
-    MOVE, MOVE, MOVE, MOVE, MOVE,
-    MOVE, MOVE, MOVE, MOVE, MOVE,
-    MOVE, MOVE,
-    HEAL, HEAL, HEAL, HEAL, HEAL,
-  ]
-}
-
-
-const noDefence3TowerAttackerBoost: MineralBoostConstant[] = [
-  RESOURCE_ZYNTHIUM_OXIDE,
-  RESOURCE_LEMERGIUM_ALKALIDE,
-  RESOURCE_KEANIUM_ALKALIDE,
-  RESOURCE_CATALYZED_GHODIUM_ALKALIDE,
-]
-
-const tier3SwampHealerSpec: CreepBodySpec = {
-  roles: [CreepRole.RangedAttacker, CreepRole.Healer, CreepRole.Mover],
-  body: [
-    TOUGH, TOUGH, TOUGH, TOUGH, TOUGH,
-    TOUGH,
-    MOVE, MOVE, MOVE, MOVE, MOVE,
-    MOVE,
-    RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK,
-    RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK,
-    RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK,
-    RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK,
-    RANGED_ATTACK, RANGED_ATTACK, RANGED_ATTACK,
-    MOVE, MOVE, MOVE, MOVE, MOVE,
-    MOVE, MOVE, MOVE, MOVE,
-    HEAL, HEAL, HEAL, HEAL, HEAL,
-    HEAL,
-  ]
-}
-const tier3SwampDismantlerSpec: CreepBodySpec = {
-  roles: [CreepRole.Worker, CreepRole.Mover],
-  body: [
-    TOUGH, TOUGH, TOUGH, TOUGH, TOUGH,
-    TOUGH,
-    MOVE, MOVE, MOVE, MOVE, MOVE,
-    MOVE,
-    WORK, WORK, WORK, WORK, WORK,
-    WORK, WORK, WORK, WORK, WORK,
-    WORK, WORK, WORK, WORK, WORK,
-    WORK, WORK, WORK, WORK, WORK,
-    WORK, WORK, WORK, WORK, WORK,
-    WORK, WORK, WORK, WORK,
-    MOVE, MOVE, MOVE, MOVE, MOVE,
-    MOVE, MOVE, MOVE, MOVE,
-  ]
-}
-
-// XZHO2,XLHO2,KHO2,XGHO2,XZH2O
-const tier3DismantlerBoost1: MineralBoostConstant[] = [
-  RESOURCE_CATALYZED_ZYNTHIUM_ALKALIDE,
-  RESOURCE_CATALYZED_LEMERGIUM_ALKALIDE,
-  RESOURCE_KEANIUM_ALKALIDE,
-  RESOURCE_CATALYZED_GHODIUM_ALKALIDE,
-  RESOURCE_CATALYZED_ZYNTHIUM_ACID,
-]
 
 export interface Season1673282SpecializedQuadProcessState extends ProcessState {
   /** parent room name */
@@ -210,45 +28,19 @@ export interface Season1673282SpecializedQuadProcessState extends ProcessState {
   targetRoomName: RoomName
   waypoints: RoomName[]
   predefinedTargetIds: Id<AttackTarget>[]
-  creepType: Season1673282SpecializedQuadProcessCreepType
+  quadType: QuadType
   creepNames: CreepName[]
   quadState: QuadState | null
 }
 
-// test tier0-d100-attacker
-// Game.io("launch -l Season1673282SpecializedQuadProcess room_name=W14S28 target_room_name=W14S26 waypoints=W14S27 creep_type=tier0-d100-attacker targets=")
-
-// W11S23 tier0-d450
-// Game.io("launch -l Season1673282SpecializedQuadProcess room_name=W9S24 target_room_name=W11S23 waypoints=W10S24,W11S24 creep_type=tier0-d450 targets=")
-// tire0-d360-dismantler
-// Game.io("launch -l Season1673282SpecializedQuadProcess room_name=W9S24 target_room_name=W11S23 waypoints=W10S24,W11S24 creep_type=tire0-d360-dismantler targets=610ea82a56c8192c44e7de28,60fc574ea19fa5e4ddb0b833")
-
-// W11S25
+// W11S23
 // tier0-d450
-// Game.io("launch -l Season1673282SpecializedQuadProcess room_name=W9S24 target_room_name=W11S25 waypoints=W10S24,W10S25 creep_type=tier0-d450 targets=611362889c09be8c65e9e5c6")
-
-// W13S27 tier0-d450
-// Game.io("launch -l Season1673282SpecializedQuadProcess room_name=W14S28 target_room_name=W13S27 waypoints=W14S30,W12S30,W12S28,W13S28 creep_type=tier0-d450 targets=")
-// Game.io("launch -l Season1673282SpecializedQuadProcess room_name=W14S28 target_room_name=W13S27 waypoints=W14S30,W12S30,W12S28,W13S28 creep_type=tire0-d360-dismantler targets=")
-// W13S27 tier3-d2000-dismantler-swamp
-// Game.io("launch -l Season1673282SpecializedQuadProcess room_name=W14S28 target_room_name=W13S27 waypoints=W14S30,W12S30,W12S28,W13S28 creep_type=tier3-d2000-dismantler-swamp targets=")
-
-// W2S29
-// Game.io("launch -l Season1673282SpecializedQuadProcess room_name=W3S24 target_room_name=W2S29 waypoints=W3S25,W2S25,W2S28 creep_type=tier0-d450 targets=")
-
-// W27S29 tire0-d360-dismantler
-// Game.io("launch -l Season1673282SpecializedQuadProcess room_name=W24S29 target_room_name=W27S29 waypoints=W24S30,W27S30 creep_type=tire0-d360-dismantler targets=")
-
-// E3S29 no-defence-3tower
-// Game.io("launch -l Season1673282SpecializedQuadProcess room_name=W6S29 target_room_name=E3S29 waypoints=W6S30,E4S30,E4S29 creep_type=no-defence-3tower targets=")
-
-// E5S29 invader-core-attacker
-// Game.io("launch -l Season1673282SpecializedQuadProcess room_name=W6S29 target_room_name=E5S29 waypoints=W6S30,E5S30 creep_type=invader-core-attacker targets=")
+// Game.io("launch -l Season1673282SpecializedQuadProcess room_name=W9S24 target_room_name=W11S23 waypoints=W10S24,W11S24 quad_type=tier0-d450 targets=")
 export class Season1673282SpecializedQuadProcess implements Process, Procedural, MessageObserver {
   public readonly identifier: string
   private readonly codename: string
 
-  private readonly boosts: MineralBoostConstant[]
+  private readonly quadSpec: QuadSpec
 
   private constructor(
     public readonly launchTime: number,
@@ -257,31 +49,14 @@ export class Season1673282SpecializedQuadProcess implements Process, Procedural,
     public readonly targetRoomName: RoomName,
     public readonly waypoints: RoomName[],
     private readonly predefinedTargetIds: Id<AttackTarget>[],
-    private readonly creepType: Season1673282SpecializedQuadProcessCreepType,
+    private readonly quadType: QuadType,
     private readonly creepNames: CreepName[],
     private quadState: QuadState | null,
   ) {
     this.identifier = `${this.constructor.name}_${this.launchTime}_${this.parentRoomName}`
     this.codename = generateCodename(this.identifier, this.launchTime)
 
-    switch (this.creepType) {
-    case "test-dismantler":
-    case "test-attacker":
-      this.boosts = [...noBoosts]
-      break
-    case "invader-core-attacker":
-    case "tier0-d100-attacker":
-    case "tier0-d450":
-    case "tire0-d360-dismantler":
-      this.boosts = [...noBoosts]
-      break
-    case "tier3-d2000-dismantler-swamp":
-      this.boosts = [...tier3DismantlerBoost1]
-      break
-    case "no-defence-3tower":
-      this.boosts = [...noDefence3TowerAttackerBoost]
-      break
-    }
+    this.quadSpec = new QuadSpec(this.quadType)
   }
 
   public encode(): Season1673282SpecializedQuadProcessState {
@@ -293,18 +68,18 @@ export class Season1673282SpecializedQuadProcess implements Process, Procedural,
       targetRoomName: this.targetRoomName,
       waypoints: this.waypoints,
       predefinedTargetIds: this.predefinedTargetIds,
-      creepType: this.creepType,
+      quadType: this.quadType,
       creepNames: this.creepNames,
       quadState: this.quadState,
     }
   }
 
   public static decode(state: Season1673282SpecializedQuadProcessState): Season1673282SpecializedQuadProcess {
-    return new Season1673282SpecializedQuadProcess(state.l, state.i, state.p, state.targetRoomName, state.waypoints, state.predefinedTargetIds, state.creepType, state.creepNames, state.quadState)
+    return new Season1673282SpecializedQuadProcess(state.l, state.i, state.p, state.targetRoomName, state.waypoints, state.predefinedTargetIds, state.quadType, state.creepNames, state.quadState)
   }
 
-  public static create(processId: ProcessId, parentRoomName: RoomName, targetRoomName: RoomName, waypoints: RoomName[], predefinedTargetIds: Id<AttackTarget>[], creepType: Season1673282SpecializedQuadProcessCreepType): Season1673282SpecializedQuadProcess {
-    return new Season1673282SpecializedQuadProcess(Game.time, processId, parentRoomName, targetRoomName, waypoints, predefinedTargetIds, creepType, [], null)
+  public static create(processId: ProcessId, parentRoomName: RoomName, targetRoomName: RoomName, waypoints: RoomName[], predefinedTargetIds: Id<AttackTarget>[], quadType: QuadType): Season1673282SpecializedQuadProcess {
+    return new Season1673282SpecializedQuadProcess(Game.time, processId, parentRoomName, targetRoomName, waypoints, predefinedTargetIds, quadType, [], null)
   }
 
   public processShortDescription(): string {
@@ -359,7 +134,7 @@ export class Season1673282SpecializedQuadProcess implements Process, Procedural,
     })
 
     const priority = ((): CreepSpawnRequestPriority => {
-      if (this.creepType === "test-dismantler") {
+      if (this.quadType === "test-dismantler") {
         return CreepSpawnRequestPriority.High
       }
       if (quad == null) {
@@ -367,68 +142,11 @@ export class Season1673282SpecializedQuadProcess implements Process, Procedural,
       }
       return CreepSpawnRequestPriority.High
     })()
-    const squadCreepCount = ((): number => {
-      switch (this.creepType) {
-      case "test-dismantler":
-      case "test-attacker":
-        return 4
-      case "invader-core-attacker":
-        return 3
-      case "tier0-d100-attacker":
-      case "tier0-d450":
-      case "tire0-d360-dismantler":
-      case "tier3-d2000-dismantler-swamp":
-      case "no-defence-3tower":
-        return 4
-      }
-    })()
+    const squadCreepCount = this.quadSpec.creepCount()
     const creepInsufficiency = squadCreepCount - this.creepNames.length
     if (creepInsufficiency > 0) {
-      const spec = ((): CreepBodySpec => {
-        switch (this.creepType) {
-        case "test-dismantler":
-          if (creepInsufficiency <= 1) {
-            return testDismantlerSpec
-          } else {
-            return testHealerSpec
-          }
-        case "test-attacker":
-          if (creepInsufficiency <= 1) {
-            return testAttackerSpec
-          } else {
-            return testHealerSpec
-          }
-        case "invader-core-attacker":
-          if (creepInsufficiency >= 2) {
-            return invaderCoreAttackerAttacker
-          } else {
-            return invaderCoreAttackerHealer
-          }
-        case "tier0-d100-attacker":
-          if (creepInsufficiency <= 1) {
-            return tier0AttackerSpec
-          } else {
-            return tier0h3HealerSpec
-          }
-        case "tier0-d450":
-          return tire0h10HealerSpec
-        case "tire0-d360-dismantler":
-          if (creepInsufficiency <= 1) {
-            return tire0DismantlerSpec
-          } else {
-            return tire0h10HealerSpec
-          }
-        case "tier3-d2000-dismantler-swamp":
-          if (creepInsufficiency <= 1) {
-            return tier3SwampDismantlerSpec
-          } else {
-            return tier3SwampHealerSpec
-          }
-        case "no-defence-3tower":
-          return noDefence3TowerAttackerSpec
-        }
-      })()
-      this.requestCreep(priority, creepInsufficiency, spec)
+      const spec = this.quadSpec.creepSpecFor(creepInsufficiency)
+      this.requestCreep(priority, creepInsufficiency, spec.roles, spec.body)
     }
 
     if (quad != null) {
@@ -459,13 +177,13 @@ export class Season1673282SpecializedQuadProcess implements Process, Procedural,
     }
   }
 
-  private requestCreep(priority: CreepSpawnRequestPriority, numberOfCreeps: number, spec: CreepBodySpec): void {
+  private requestCreep(priority: CreepSpawnRequestPriority, numberOfCreeps: number, roles: CreepRole[], body: BodyPartConstant[]): void {
     World.resourcePools.addSpawnCreepRequest(this.parentRoomName, {
       priority,
       numberOfCreeps,
       codename: this.codename,
-      roles: spec.roles,
-      body: spec.body,
+      roles,
+      body,
       initialTask: null,
       taskIdentifier: this.identifier,
       parentRoomName: null,
@@ -657,10 +375,11 @@ export class Season1673282SpecializedQuadProcess implements Process, Procedural,
   }
 
   private isBoosted(creep: Creep): boolean {
-    if (this.boosts.length <= 0) {
+    const boosts = this.quadSpec.boosts
+    if (boosts.length <= 0) {
       return true
     }
-    return this.boosts.every(boost => {
+    return boosts.every(boost => {
       const boostablePart = boostableCreepBody(boost)
       if (creep.getActiveBodyparts(boostablePart) <= 0) {
         return true
@@ -675,7 +394,7 @@ export class Season1673282SpecializedQuadProcess implements Process, Procedural,
   }
 
   private boostCreep(creep: Creep): void {
-    const unboostedType = this.boosts.find(boost => {
+    const unboostedType = this.quadSpec.boosts.find(boost => {
       return creep.body.some(body => {
         if (body.boost != null) {
           return false
